Move wrap(), wrap_axesplot() to tfplot.wrapper
